requirements.txt
Flask
google-generativeai
google-cloud-aiplatform
googlemaps
python-dotenv
